lib: give a way to close the responses returned by events.stream()
events.stream() is retrieving a streaming HTTP response and returning
iterator based on that. Once the iterator is no longer used, the
response is not explicitly closed.

With this change, the iterator returned gets a close() method and
implements the context manager protocol, allowing responses to be
closed once the iterator is no longer needed.